api/service/requirements.txt
aerich==0.6.2
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.0.3
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.73.0
h11==0.13.0
idna==3.3
iso8601==0.1.16
pydantic==1.9.0
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-decouple==3.5
pytz==2021.3
sniffio==1.2.0
starlette==0.17.1
tomlkit==0.9.0
tortoise-orm==0.18.1
typing_extensions==4.0.1
uvicorn==0.17.1
